requirements.txt
rembg
Pillow
PyInstaller
